Fix UnicodeDecodeError by removing pexcept dependency
The pexept module can sometimes cause errors when running
through the Ansible builtin except module
(https://github.com/ansible/ansible/issues/29351).

To prevent these issues from occuring and messing with the CI,
we remove the except module call and replace it with a shell
call to except directly. This also removes the pexcept package
as a dependency from ca_server.