style: added gitlint integration for commit linting
Currently, we use commitlint to check commit messages
in the CI pipeline. This is far from optimal, as it
means that there's no way to check commit messages
before pushing (unless we want everyone to install npm
just to lint some commit messages).

This commit replaces that with gitlint, a commit linter
that supports conventional commits as used by this project
and with pre-commit integration, making it easy to setup
and integrate.