feat: add arm64 support to step_ca and step_cli
This commit adds support for alternative architectures
(arm64 for now) to both installer roles.

As the debian packages only come with an amd64 executable,
I am dropping support for installing step via deb packages.
This should be transparent to any user running this role
as the tarball installation uses the exact same paths