feat: use root user for step-cli bootstrapping
BREAKING CHANGE

This commit removes the 'step' service user from this collection,
replacing it with the root user. The main motivations behind this change is
to make calling components of this collection and step-cli itself easier.

Previously, the docs suggested that to call step-cli via a module or on
the CLI, you'd always need to use the step service user. This wasn't really
true as you could always use the --ca-url and --root flas (or --ca-config
if local) instead, but this was discouraged for no good reason. This commit
clears up the documentation a bit and clearly shows the user what options
they have.

Starting from now on, users can use either the parameters above or the root
user to execute their modules. This makes playbooks easier to write as there
is no need for become_user flags anymore. It also helps with issues that can
arise when becoming an inprivileged user, such as the need for world readable
tmpfiles.

The service user was also used to run the step renewal job in step_acme_cert -
this job is now run by the root user. While marginally less secure, it removes
a lot of extra required work from our roles, such as the sudo permissions for
the step user.

Existing setups should continue to function for the most part. Since this
change doesn't actually alter the existing step user, it shouldn't break it.
That said, there might be other unforseen consequences and it is reccomended
that you adjust your playbooks to this new method.