ci: fix tags not pointing to release commits (#70)
Until now, the release tags always pointed to the last commit before
the actual release (the commit including the galaxy.yml and CHANGELOG.md)
update, apparently due to the create-release action using the last commithash
at the time of checkout, which happens at the beginning of the workflow,
thus ignoring the new release commit. This commit fixes that behavior
by explicitly setting the commit hash to the hash of the release commit.