fix(step_bootstrap_host): support STEPPATH ENV VAR
This amends the checks and installation of STEP CA Root certificate
to use STEPPATH if set.  If not set, will mimic the `step-cli`
behavior and default to `$HOME/.step`.

Note that the previous implementation utilized the ansible
`lookup(env, 'STEPPATH')` filter, but that will not work for
environment variables set via the `environment:` block,
per the documentation:

https://docs.ansible.com/ansible/latest/user_guide/playbooks_environment.html

> The environment: keyword does not affect Ansible itself, Ansible
> configuration settings, the environment for other users, or the
> execution of other plugins like lookups and filters.

What this does allow is setting of the environment variable at a play
or task level; e.g.:

```yaml
- hosts: step_client
  become: yes
  environment:
    STEPPATH: "{{ sift_steppath }}"
  tasks:
   [...]
```