fix: eliminate "expected errors" when looking for step-cli/certs
When looking for the step-cli at the begging of a role,
we can use the "command" shell builtin to check for the binaries presence
instead of calling it directly.

This has the advantage of elminating the ugly "failed" task that would
normally occur when running the step_cli or step_bootstrap_hosts roles
for the first time. For consistencys sake, we also use this new method
in roles like step_acme_cert that error out upon not finding step-cli.

We can use the same pattern to also catch the error thrown by step-cli
upon finding an invalid/not-yet-installed ca cert in step_bootstrap_host.

Fixes https://github.com/maxhoesel/ansible-collection-smallstep/issues/96,