make step-cli/ca install dir user-selectable
With this change, we now install step-cli/ca to the dir defined by the user.

This also fixes a potential issue where if a user specified a custom
executable via `step_bootstrap_host` that did not exist,
the role would have just installed step-cli to the default directory
instead of `step_cli_executable`.

From now on, we will determine the installation for step-cli path as follows:
- Use `step_cli_executable` if it is a path
- Use `step_cli_install_dir` + `step_cli_executable` if the latter is a filename

For step-ca, we simply require `step_ca_executable` to be an absolute path.
This is not possible for step-cli without breaking backwards compatibility as the variable
could also be a filename.