replace `except` with --password-file flags
This change removes the need to shell out to `except`
when initializing the CA and setting up password files.
Newer versions of the step crypto change-pass command
include an option to pass the password via a file instead of interactive input.
As this removes the need for `except`, we should use that.