Move molecule tests to podman
As discussed in #105, molecule-docker is no longer a suitable driver
for our needs (specifically due to systemd support), so it's about
time we switched to a container runtime that actually supports systemd.

With the 4.0 release, podman is finally up to the task of running our
integration tests. This change migrates molecule to podman
and updates the contributing documentation to account for the changes.