step_ca_provisioner: make idempotent, remove 'updated' state
This commit allows `step_ca_provisioner` to determine whether
a given provisioner has been changed, be that through creation
or an update.

We do this by comparing the provisioner config before and after
running the `create/update` command.

This has the side-effect of making this module check-mode-incompatible,
since there is no way to tell whether a provisioner has changed without
actually applying it.

This commit also fixes some behavior around the --create parameter
to ensure idempotency,
as well as some other minor details.