modules: refactor cli_wrapper to Command-based approach
In preparation for addressing #352, this commit changes the
workflow for executing step-cli commands into a builder pattern:

Modules now first construct a command with all parameters already
passed in, before then executing it with a single invocation.

This is less error-prone than the old style of moving around raw dicts
created with build_params() and allows us to extend the abstraction
with new features in the future.
For now, only the new tmpfile is included, which will allow modules
to easily pass in strings such as passwords through temporary files.