modules: add `_password` args as alternative to password_files
this PR adds a `_password` parameter to go along with
every currently existing `password_file` module parameter.
As discussed in #352, this makes passing in passwords to tasks
much more convenient, as users no longer need to create and delete
password files manually.

To ensure no sensitive password files remain on the system, they
are created in a temporary directory that is handled through a context manager.