fix CI builds and rework testing harness (#419)
* lock transitive python dependencies with pip-tools

After yet another breaking package update
(https://github.com/docker/docker-py/pull/3257),
its high time we permanently pin all development packages
*and their dependencies* to a known.-good version.
pip-tools lets us do that without losing the simple requirements.txt
file that allows easy installation, so lets use that.

* rework testing harness for py12, remove pytest-virtualenv dep

a little while ago, we were starting to get "imp" import errors
in CI builds, which I initially were assumed to be from pytest.
In fact, it was the pytest-virtualenv package that we used
to generate a virtualenv with a specific ansible version for tests.

As there is no easy replacement for this package, and the package
itself is stale (no release in 5 years), the testing harness now just
installs the desired version of ansible-core straight into the main
devel virtualenv. This is a bit ugly as it could interfere with other
user tasks, but it leaves the interface the same and helps simplify the test fixture code.
Said code also now reliably isolates tes environments from another and the user-wide collections.

Combined with package pinning introduced in the previous commit,
we should no longer have CI failures due to packages or their deps
updating and breaking things.